Create Temperature Converter.m
Write a MATLAB program that converts temperature from Celsius to Fahrenheit or from
Fahrenheit to Celsius based on user input.
i. Prompt the user to enter the temperature value and the unit of the temperature (Celsius
or Fahrenheit).
ii. Based on the user's input, perform the appropriate conversion using the following
formulas:
 Celsius to Fahrenheit: Fahrenheit = (Celsius * 9/5) + 32
 Fahrenheit to Celsius: Celsius = (Fahrenheit - 32) * 5/9
iii. Display the converted temperature to the user.